Admin/requirements.txt
asgiref==3.11.0
Django==4.2.27
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1
PyMySQL==1.1.2
sqlparse==0.5.5
tzdata==2025.3